CLN: moved Envisage Plugin subclasses to core_plugins directory
